Make 3D leaf render surface decision compatible with TransformInterop.
The decision about when to create a render surface at the leafward edges
of preserve-3d scenes was not compatible with the changes to preserve-3d
made in TransformInterop in a number of ways.

With TransformInterop, being the leaf of a 3D scene (where we define
leaf as the element *all* of whose descendants are flattened into it)
does not require any unusual properties on the element; it merely
requires that the parent have transform-style: preserve-3d but the
element not do so.  (Prior to TransformInterop, it required something
that created a node in the transform tree; see NeedsTransform().)

This requires that we do two things:

 * ensure that we create a PaintLayer for such elements, and

 * make the decision about whether an element is in this flattening
   position in the effect node in the paint property tree while we're
   still operating on layout objects and have accurate information about
   element parent/child relationships.

This also avoids a case in the old code (which I'm not convinced was
bug-free in this respect prior to TransformInterop) where the code could
get set state on an effect node when operating on a transform node that
was a descendant of that effect node's transform node (but one that
shared the same effect node).

Note that the existing web test
compositing/overflow-trumps-transform-style.html is the only test that
fails if this code to cause creation of a render surface is removed.

The newly-added tests preserve3d-and-flattening-00{1,2,3}.html fail
prior to this change.   001 is the one that is closest to the broken
site; it does not fail if the code to create a render surface is just
removed.  002 and 003, however, also fail if the code is removed.  (The
distinction between 002 and 003 was needed to get more complete test
coverage with earlier versions of this change, but the revisions needed
to fix both of them mean the distinction is no longer as relevant, but I
kept them both anyway.)

Fixed: 1255544
Change-Id: I4d8ae4c6a134cc1605188779760079be1e9c7f90