Clamp <length> in AccumulateLengthArray and add DCHECK in Blend()
According to spec [1], the infinity and NaN should be clamped before other components consume the values.

But The infinity and NaN values could escape calc()  from AccumulateLengthArray used for interpolation.

Therefore, Add ClampLength in AccumulateLengthArray to prevent it and add DCHECK(infinite) statement in Blend function to ensure and check the no infinity and NaN values.

The Spec: [1]
Design docs: [2]

[1]  https://drafts.csswg.org/css-values/#calc-type-checking
[2] https://bit.ly/349gXjq

Bug: 1133390
Change-Id: Ibb7dbee997d883968c855754ab93f184c3f09306