Fix interactionId when click gets the wrong PointerId
Since click events sometimes do not get the correct PointerId, we use
the last PointerId of a pointerdown or pointerup seen. But we also
take care to reset this last seen when we see other kinds of events,
to prevent the click from being accidentally attributed to an
interaction in the past that did not trigger it.

Bug: 1264930
Change-Id: Ieb5773f3a6c9a9cfd46b428b5a925c4535060386