[GridFragmentation] Adjust grid-geometry/etc from an item expansion
When a grid-item expands due to fragmentation, we need to adjust the
grid-geometry, intrinsic-block-size, and subsequent offsets.

To do this we record all the item expansions, in that row, and adjust by
the maximum.

We re-run the inner "PlaceItems" with the new grid-geometry such that
subsequent fragments etc are shifted as appropriate.

Bug: 614667
Change-Id: Ia66da118251c568dbee84f16aa32313bf391c280